BM-992: cap the min batch size to max concurrent proofs (#696)
Realized we have a small footgun here where if the min batch size is
higher than the max concurrent proofs, the prover will just wait until
the end of expiry to aggregate and submit. Warn log to nudge operator to
update config but also capping at max concurrent proofs

+++ crates/broker/src/config.rs
@@ -252,7 +252,7 @@ pub struct BatcherConfig {
     pub batch_max_time: Option<u64>,
     /// Batch size (in proofs) before publishing
     #[serde(alias = "batch_size")]
-    pub min_batch_size: Option<u64>,
+    pub min_batch_size: Option<u32>,
     /// Max combined journal size (in bytes) that once exceeded will trigger a publish
     #[serde(default = "defaults::batch_max_journal_bytes")]
     pub batch_max_journal_bytes: usize,
